In Sysman fix C++ programming warnings
Fix warnings related to unused/uninitialized variables etc.

+++ level_zero/tools/source/sysman/linux/pmu/pmu.h
@@ -13,7 +13,7 @@ class LinuxSysmanImp;
 class PmuInterface {
   public:
     virtual ~PmuInterface() = default;
-    virtual int64_t pmuInterfaceOpen(uint64_t config, int group, uint64_t format) = 0;
+    virtual int64_t pmuInterfaceOpen(uint64_t config, int group, uint32_t format) = 0;
     virtual int pmuReadSingle(int fd, uint64_t *data, ssize_t sizeOfdata) = 0;
     static PmuInterface *create(LinuxSysmanImp *pLinuxSysmanImp);
 };
